[auxdata] move icechart & rio code to auxdata submodule
