--- README.rst
@@ -1,6 +1,0 @@
-pysiral: python sea ice radar altimetry processing library
-==========================================================
-
-test
-test2
-
